+++ src/strategies/erc721-with-tokenid-range-weights-simple/examples.json
@@ -1,0 +1,23 @@
+[
+  {
+    "name": "Example query",
+    "strategy": {
+      "name": "erc721-with-tokenid-range-weights-simple",
+      "params": {
+        "address": "0x696115768bbef67be8bd408d760332a7efbee92d",
+        "symbol": "LINKSDAO",
+        "tokenIdWeightRanges": [
+          { "start": 1, "end": 100, "weight": 1 },
+          { "start": 6364, "end": 6464, "weight": 4 }
+        ]
+      }
+    },
+    "network": "1",
+    "addresses": [
+      "0x7d863d74917191685616217c8ab1a77e73e79f21",
+      "0xb111dabb8edd8260b5c1e471945a62be2ee24470",
+      "0x285bd3a5c9cae024d70c7e70fcfc59db03637549"
+    ],
+    "snapshot": 13995858
+  }
+]
